% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibraryIntegrationApi.r
\name{LibraryIntegrationApi_get_parents_by_study}
\alias{LibraryIntegrationApi_get_parents_by_study}
\title{Retrieve group metadata by querying study ID (accession)}
\usage{
LibraryIntegrationApi_get_parents_by_study(id, ...)
}
\arguments{
\item{id}{Unique identifier (accession) of the object.}
}
\description{
Retrieve group metadata by querying study ID (accession)
}
