% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkageApi.r
\name{LinkageApi_get_data_types_links}
\alias{LinkageApi_get_data_types_links}
\title{List all possible links between data types that match the specified criteria.}
\usage{
LinkageApi_get_data_types_links(type, ...)
}
\arguments{
\item{type}{Return only links with the specified data type.}
}
\description{
List all possible links between data types that match the specified criteria.
}
