% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkageApi.r
\name{LinkageApi_get_links_by_ids}
\alias{LinkageApi_get_links_by_ids}
\title{Finds existing links by passing many IDs.  Pagination goes through all links matched the criteria.}
\usage{
LinkageApi_get_links_by_ids(request, ...)
}
\arguments{
\item{request}{}
}
\description{
Finds existing links by passing many IDs.  Pagination goes through all links matched the criteria.
}
