% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkageApi.r
\name{LinkageApi_get_links_by_params}
\alias{LinkageApi_get_links_by_params}
\title{Finds existing links matching the specified criteria.}
\usage{
LinkageApi_get_links_by_params(
  first_id,
  first_type,
  second_id,
  second_type,
  offset,
  limit,
  ...
)
}
\arguments{
\item{first_id}{Object ID (accession) (e.g. accession of study)}

\item{first_type}{Type of the object (e.g. study)}

\item{second_id}{Object ID (accession) (e.g. accession of study)}

\item{second_type}{Type of the object (e.g. study)}

\item{offset}{Param says to skip that many links before beginning to return links.}

\item{limit}{Param says to limit the count of returned links.}
}
\description{
Finds existing links matching the specified criteria.
}
