% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreparationIntegrationApi.r
\name{PreparationIntegrationApi_get_parents_by_study}
\alias{PreparationIntegrationApi_get_parents_by_study}
\title{Retrieve group metadata by querying study ID (accession)}
\usage{
PreparationIntegrationApi_get_parents_by_study(id, ...)
}
\arguments{
\item{id}{Unique identifier (accession) of the object.}
}
\description{
Retrieve group metadata by querying study ID (accession)
}
