% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariantIntegrationApi.r
\name{VariantIntegrationApi_get_parents_by_study}
\alias{VariantIntegrationApi_get_parents_by_study}
\title{Retrieve group metadata by querying study ID (accession)}
\usage{
VariantIntegrationApi_get_parents_by_study(
  id,
  use_versions,
  returned_metadata_fields,
  ...
)
}
\arguments{
\item{id}{Unique identifier (accession) of the object.}

\item{use_versions}{Specify which versions of omics data files are used in the query. By default the active version is used. See Versioning above. Syntax:  \\* or \code{v<version number>} or \code{<CHAIN_ID>}:\code{v<version number>} or \code{<CHAIN_ID>}:\code{<accession1,accession2,..>}}

\item{returned_metadata_fields}{The parameter defines amount of metadata attributes to return:  1. \code{template} - return metadata attributes according to applied template, if object doesn’t have applied template default template will be used. This is the default for User endpoints. 2. \code{all} - return all metadata attributes with values and null attributes, if they are present in the applied template. This is the default for Curator endpoints.}
}
\description{
Retrieve group metadata by querying study ID (accession)
}
